Fix error on import of external modules
There were import errors for the fp-ts package after the change in server/tsconfig.json from "module": "commonjs" to "module": "ESNext" in #78655b17.
The error basically asked that the changes in this commite were made in tsconfig.json (moduleResultion -> node) and in package.json("type": "module").
Also an import (fp-ts/Json) which worked with commonjs did not work now, so it was also fixed.