Fix operation of sim due to deepcopy issues
the deepcopy of classes in TS loses class info (a class is converted to an Object).
There are multiple ways of fixing this, but the simplest in this case was to change all instanceof checks to x.type_name === "ClassName" checks and overrride some type chaecking.
The simulation is now working correctly.